Fix InverseMedianRank

+++ src/pykeen/metrics/ranking.py
@@ -1351,7 +1351,7 @@ class InverseMedianRank(RankBasedMetric):
     def __call__(
         self, ranks: np.ndarray, num_candidates: np.ndarray | None = None, weights: np.ndarray | None = None
     ) -> float:  # noqa: D102
-        return np.reciprocal(weighted_median(a=ranks, weights=weights)).item()
+        return np.reciprocal(weighted_median(a=ranks, weights=weights), dtype=float).item()
 
 
 @parse_docdata
